fix: warnings

+++ src/semantic/chocopy_type.cpp
@@ -54,7 +54,7 @@ bool FunctionDefType::operator==(const FunctionDefType &f2) const {
     auto &a = this->params;
     auto &b = f2.params;
     if (a.size() == b.size()) {
-        for (int i = 1; i < a.size(); i++) {
+        for (size_t i = 1; i < a.size(); i++) {
             if (a.at(i)->get_name() != b.at(i)->get_name()) return false;
         }
         if (this->return_type->get_name() != f2.return_type->get_name())
